CODE RUB: Switch logic to appropriate file
