+++ ISL.ReIdentification.WebApps.Portals/ISL.ReIdentification.Portals.Server.Tests.Acceptance/Apis/ReIdentificationApiTests.Logic.ExpireRenewImpersonationContextTokens.cs
@@ -2,43 +2,36 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
+using System;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Force.DeepCloner;
-using ISL.ReIdentification.Core.Models.Coordinations.Identifications;
+using ISL.ReIdentification.Portals.Server.Tests.Acceptance.Models.Accesses;
 using ISL.ReIdentification.Portals.Server.Tests.Acceptance.Models.ImpersonationContexts;
 
 namespace ISL.ReIdentification.Portals.Server.Tests.Acceptance.Apis
 {
     public partial class ReIdentificationApiTests
     {
-        [Fact]
-        public async Task ShouldUpdateApprovalStatusOnImpersonationContextApprovalAsync()
+        [Fact(Skip = "Access to Storage keeps failing")]
+        public async Task ShouldExpireRenewImpersonationContextTokensAsync()
         {
             // given
             ImpersonationContext randomImpersonationContext = CreateRandomImpersonationContext();
             ImpersonationContext existingImpersonationContext = randomImpersonationContext.DeepClone();
-            bool isApproved = true;
-            existingImpersonationContext.IsApproved = !isApproved;
-            existingImpersonationContext.ResponsiblePersonEntraUserId = TestAuthHandler.SecurityOid;
+            existingImpersonationContext.IsApproved = false;
             await this.apiBroker.PostImpersonationContextAsync(existingImpersonationContext);
-
-            ApprovalRequest inputApprovalRequest = new ApprovalRequest
-            {
-                ImpersonationContextId = existingImpersonationContext.Id,
-                IsApproved = isApproved
-            };
+            Guid inputImpersonationContextId = existingImpersonationContext.Id;
 
             // when
-            await this.apiBroker.PostImpersonationContextApprovalAsync(inputApprovalRequest);
+            AccessRequest actualAccessRequest =
+                await this.apiBroker.PostImpersonationContextGenerateTokensAsync(inputImpersonationContextId);
 
             // then
-            ImpersonationContext retrievedImpersonationContext =
-                await this.apiBroker.GetImpersonationContextByIdAsync(existingImpersonationContext.Id);
-
-            retrievedImpersonationContext.IsApproved.Should().Be(isApproved);
-
-            await this.apiBroker.DeleteImpersonationContextByIdAsync(retrievedImpersonationContext.Id);
+            actualAccessRequest.ImpersonationContext.InboxSasToken.Should().NotBeNullOrEmpty();
+            actualAccessRequest.ImpersonationContext.OutboxSasToken.Should().NotBeNullOrEmpty();
+            actualAccessRequest.ImpersonationContext.ErrorsSasToken.Should().NotBeNullOrEmpty();
+            await this.apiBroker.DeleteImpersonationContextByIdAsync(actualAccessRequest.ImpersonationContext.Id);
         }
     }
 }
